web: split sparkline helpers into their own module for fast refresh
